feat: add classes to delegate editor of table cells for mapping table edition
